requirements.txt
pdfplumber
pdf2image
pytesseract
anthropic
python-dotenv
pyyaml
pytest
